requirements.txt

Flask==3.0.3
python-dotenv==1.0.1
requests==2.32.3
SQLAlchemy==2.0.36
pydantic==2.9.2
google-api-python-client==2.146.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-generativeai==0.7.2
